Extract ImapMessage from ImapStore
