+++ k9mail-library/src/main/java/com/fsck/k9/mail/store/imap/FetchBodyCallback.java
@@ -7,7 +7,6 @@ import java.util.Map;
 import com.fsck.k9.mail.Message;
 import com.fsck.k9.mail.MessagingException;
 import com.fsck.k9.mail.filter.FixedLengthInputStream;
-import com.fsck.k9.mail.store.imap.ImapStore.ImapMessage;
 
 
 class FetchBodyCallback implements ImapResponseCallback {
